requirement.txt
altgraph==0.17.3
anaconda==0.0.1.1
aniso8601==10.0.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dash==3.0.0
dash-bootstrap-components==2.0.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
disk-smart-tool==0.0.1
dominate==2.9.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.0
fqdn==1.5.1
future==0.18.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
narwhals==1.31.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pefile==2022.5.30
pillow==10.2.0
platformdirs==4.3.6
plotly==6.0.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.2.0
pyreadline3==3.5.4
pySMART==1.4.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-json-logger==2.0.7
python-socketio==5.12.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.0
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
statsmodels==0.14.4
stringcase==1.2.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tk==0.1.0
tomli==2.1.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
visitor==0.1.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
WMI==1.5.1
wsproto==1.2.0
WTForms==3.2.1
xlrd==2.0.1
zipp==3.21.0
